requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-mathfilters==1.0.0
mysql==0.0.2
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
